+++ etc/headers/h2m_structs.h
@@ -1,12 +1,7 @@
 // -*- coding: utf-8 -*-
 // (c) 2013-2015 Hydro2Motion Developers, h2m@hm.edu
 
-// target: Mbed
-//#include "mbed.h"
-
-// target: PC
-#include "stdint.h"
-
+#include "mbed.h"
 
 // Program variable
 //----------------------------------------------------------------------------------------------
@@ -17,6 +12,10 @@ typedef struct struct_program         // added 06.03.2014 C.L.
     struct_program()
     : length(13), ID(0)
     {}
+
+    struct_program(): length(13), ID(0) {}
+    uint8_t  length;             //  1 Length of struct (byte)
+    uint8_t  ID;                 //  2 Struct ID
     */
     uint8_t  length = 13;        //  1 Length of struct (byte)
     uint8_t  ID     =  0;        //  2 Struct ID